解决注入问题并优化查询，但尚有bug，明天继续解决

+++ func_model/del_account.c
@@ -28,18 +28,21 @@ int del_account(void)
       if(!mysql_query(&mysql,"start transaction")){
         //开始事务
         //开始删除passwd的记录
-        snprintf(dest,200,"delete from passwd where account='%s'",login_name);
+        if(DEL_ACCOUNT)
+          snprintf(dest,200,"set @var_acc='%s';execute pre_del_account using @var_acc",login_name);
+        else
+          snprintf(dest,200,"delete from passwd where account='%s'",login_name);
         if(!mysql_query(&mysql,dest)){
           //passwd记录删除成功，开始删除对应的event和profile表
-          //先删profile表
-          snprintf(dest,200,"drop table %s_profile",login_name);
+          //先删profile表，构造预编译语句
+          snprintf(dest,200,"set @var_pre_profile='%s_profile';prepare del_profile from 'drop table ?';execute del_profile using @var_pre_profile;drop prepare del_profile",login_name);
           if(!mysql_query(&mysql,dest)){
             //删除profile表成功，开始删除event表
-            snprintf(dest,200,"drop table %s_event",login_name);
+            snprintf(dest,200,"set @var_pre_event='%s_event';prepare del_event from 'drop table ?';execute del_event using @var_pre_event;drop prepare del_event",login_name);
             if(!mysql_query(&mysql,dest)){
               //都删除成功，直接返回
               if(!mysql_query(&mysql,"commit")){
-                printf("\t你还是那么狠心，按回车吧，前任");
+                printf("\t你还是那么狠心，按回车吧，哼");
                 getchar();
                 return 1;
               }
